Automated Hiring System/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
spacy
dateparser
python-multipart
requests
